Add file handlers to 'Tasks_5' programs
There are two handlers - one for Windows and one for Linux.
They are used to load/save files in the same directory as the program,
not the directory where the program is running.
Thanks to the 'define' directives, the program will use the correct
handler depending on the operating system.